+++ R/par-pattern.R
@@ -1,0 +1,10 @@
+#' Parameter Pattern
+#'
+#' @return A string of the regular expression for a parameter name.
+#' @export
+#'
+#' @examples
+#' par_pattern()
+par_pattern <- function() {
+  "[[:alpha:]][[:alnum:]._]*"
+}