+++ tests/testthat/test-par-pattern.R
@@ -1,0 +1,3 @@
+test_that("multiplication works", {
+  expect_identical(par_pattern(), "[[:alpha:]][[:alnum:]._]*")
+})