+++ NEWS.md
@@ -1,3 +1,10 @@
+# extras 0.1.0
+
+- Added `chk_indices()` and `vld_indices()` to check (validate) if an object is a list of indices ie a vector of one or more positive integer values.
+- Added `par_pattern()` to provide string of regular expression for a parameter name.
+- Added `as_list_unnamed()` generic which by default strips all attributes except names.
+- Modified `chk_pars()` and `vld_pars()` to permit missing values and duplicates.
+
 # extras 0.0.1
 
 - Initial submission to CRAN.